Use laminas/migration-tool to migrate from zend to laminas libraries. Also updated referenced towards zend to laminas
