+++ .travis.yml
@@ -5,6 +5,11 @@ dist: trusty
 php:
   - 5.5
   - 5.6
+  - 7.0
+  - 7.1
+  - 7.2
+  - 7.3
+  - 7.4
 
 matrix:
   exclude:
@@ -18,13 +23,21 @@ matrix:
       env: dependencies=highest
     - php: 7
       env: dependencies=highest
+    - php: 7.1
+      env: dependencies=highest
+    - php: 7.2
+      env: dependencies=highest
+    - php: 7.3
+      env: dependencies=highest
+    - php: 7.4
+      env: dependencies=highest
 
 before_script:
   - composer self-update
   - if [ -z "$dependencies" ]; then composer install --prefer-dist -n; fi;
   - if [ "$dependencies" = "lowest" ]; then composer update --prefer-dist --prefer-lowest -n; fi;
-  - if [ "$dependencies" = "highest" ]; then composer require zendframework/zend-mvc-console:^1.1 --prefer-dist -n; fi;
+  - if [ "$dependencies" = "highest" ]; then composer require laminas/laminas-mvc-console:^1.1 --prefer-dist -n; fi;
 
 script:
   - ./vendor/bin/phpunit
-  - ./vendor/bin/phpcs --standard=PSR2 ./src ./tests/AssetManagerTest
+  - ./vendor/bin/phpcs --standard=PSR2,PSR12 ./src ./tests/AssetManagerTest
