+++ README.md
@@ -6,7 +6,7 @@ By [Wesley Overdijk](http://blog.spoonx.nl/) and [Marco Pivetta](http://ocramius
 
 ## Introduction
 This module is intended for usage with a default directory structure of a
-[ZendSkeletonApplication](https://github.com/zendframework/ZendSkeletonApplication/). It provides functionality to load
+[LaminasSkeletonApplication](https://github.com/laminas/laminas-mvc-skeleton/). It provides functionality to load
 assets and static files from your module directories through simple configuration.
 This allows you to avoid having to copy your files over to the `public/` directory, and makes usage of assets very
 similar to what already is possible with view scripts, which can be overridden by other modules.
