+++ src/AssetManager/Controller/ConsoleControllerFactory.php
@@ -4,9 +4,9 @@ namespace AssetManager\Controller;
 
 use AssetManager\Service\AssetManager;
 use Interop\Container\ContainerInterface;
-use Zend\ServiceManager\AbstractPluginManager;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\ServiceManager\AbstractPluginManager;
+use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 
 class ConsoleControllerFactory implements FactoryInterface
 {
