+++ src/AssetManager/Resolver/CollectionResolver.php
@@ -8,7 +8,7 @@ use AssetManager\Exception;
 use AssetManager\Service\AssetFilterManager;
 use AssetManager\Service\AssetFilterManagerAwareInterface;
 use Traversable;
-use Zend\Stdlib\ArrayUtils;
+use Laminas\Stdlib\ArrayUtils;
 
 /**
  * This resolver allows the resolving of collections.
