+++ src/AssetManager/Service/AliasPathStackResolverServiceFactory.php
@@ -6,10 +6,10 @@ use AssetManager\Resolver\AliasPathStackResolver;
 use AssetManager\Resolver\PathStackResolver;
 use Interop\Container\ContainerInterface;
 use Interop\Container\Exception\ContainerException;
-use Zend\ServiceManager\Exception\ServiceNotCreatedException;
-use Zend\ServiceManager\Exception\ServiceNotFoundException;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\ServiceManager\Exception\ServiceNotCreatedException;
+use Laminas\ServiceManager\Exception\ServiceNotFoundException;
+use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 
 class AliasPathStackResolverServiceFactory implements FactoryInterface
 {
