+++ src/AssetManager/Service/AssetCacheManagerServiceFactory.php
@@ -3,8 +3,8 @@
 namespace AssetManager\Service;
 
 use Interop\Container\ContainerInterface;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 
 /**
  * Factory for the Asset Cache Manager Service
