+++ src/AssetManager/Service/CollectionResolverServiceFactory.php
@@ -4,8 +4,8 @@ namespace AssetManager\Service;
 
 use AssetManager\Resolver\CollectionResolver;
 use Interop\Container\ContainerInterface;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 
 class CollectionResolverServiceFactory implements FactoryInterface
 {
