+++ tests/AssetManagerTest/Config/ModuleServiceManagerConfigTest.php
@@ -4,8 +4,8 @@ namespace AssetManagerTest\Config;
 
 use AssetManager\Service\MimeResolver;
 use PHPUnit_Framework_TestCase;
-use Zend\ServiceManager\Config;
-use Zend\ServiceManager\ServiceManager;
+use Laminas\ServiceManager\Config;
+use Laminas\ServiceManager\ServiceManager;
 
 /**
  * Test to ensure config file is properly setup and all services are retrievable
