+++ tests/AssetManagerTest/Service/CollectionResolverServiceFactoryTest.php
@@ -5,7 +5,7 @@ namespace AssetManagerTest\Service;
 use AssetManager\Resolver\CollectionResolver;
 use AssetManager\Service\CollectionResolverServiceFactory;
 use PHPUnit_Framework_TestCase;
-use Zend\ServiceManager\ServiceManager;
+use Laminas\ServiceManager\ServiceManager;
 
 class CollectionResolverServiceFactoryTest extends PHPUnit_Framework_TestCase
 {
