+++ tests/AssetManagerTest/Service/MapResolverServiceFactoryTest.php
@@ -5,7 +5,7 @@ namespace AssetManagerTest\Service;
 use AssetManager\Resolver\MapResolver;
 use AssetManager\Service\MapResolverServiceFactory;
 use PHPUnit_Framework_TestCase;
-use Zend\ServiceManager\ServiceManager;
+use Laminas\ServiceManager\ServiceManager;
 
 class MapResolverServiceFactoryTest extends PHPUnit_Framework_TestCase
 {
