+++ tests/AssetManagerTest/Service/PrioritizedPathsResolverServiceFactoryTest.php
@@ -5,7 +5,7 @@ namespace AssetManagerTest\Service;
 use AssetManager\Resolver\PrioritizedPathsResolver;
 use AssetManager\Service\PrioritizedPathsResolverServiceFactory;
 use PHPUnit_Framework_TestCase;
-use Zend\ServiceManager\ServiceManager;
+use Laminas\ServiceManager\ServiceManager;
 
 class PrioritizedPathsResolverServiceFactoryTest extends PHPUnit_Framework_TestCase
 {
